test: ensure line assertions match local/remote
When deployed to cloud run, the `make test-acceptance` test were asserting against a specific line that was bumped by
one due to `curl` debug output.

This change changes that debug to point to FD 3 (bats' debug output) and prefixes with a hash for TAP compliance. Then it comments out the debug because we don't need it.