fix: return empty file behavior to match previous versions
The changes made meant that it could return null rather than an error.

The original behavior isn't ideal as a not particularly fatal issue of
having a single empty file in multiple provided causes the whole thing
to stop and return an error.

It'll probably take a breaking change to return the results + an error
with a log. Or just looking at some existing error output structures
from kubeval and making it similar.