feat: add fileName and IDs to the json output
Added FileName and IDs to the output. We need this for use with sarif
later. Might also be useful for other users.

We also have a Rules value that holds all unique rules that were
matched. This is excluded from the JSON output for now, we might make it
available with a flag (as it's quite verbose)

Co-authored-by: James Callaghan <james@control-plane.io>