fix: ensure json schema location can be configured
Since the migration to kubeconform, the flag `--schema-dir` does not
have any effect when provided. This commit replaces it with the flag
`--schema-location` that can be used to provide the schema remotely with
http or locallly.

This commit also adds an option to disable the schema validation. This
will be useful when we will use kubesec as a library for kubectl-kubesec and
kubesec-webhook. For these projects, we are getting the object from the
cluster so we know the object is properly formated.

Signed-off-by: ludo <controlplane@spiarh.fr>